--- models/blogpost.js
@@ -1,25 +1,0 @@
-'use strict';
-const {
-  Model
-} = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
-  class blogPost extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
-    static associate(models) {
-      // define association here
-    }
-  };
-  blogPost.init({
-    userId: DataTypes.STRING,
-    blogId: DataTypes.INTEGER,
-    text: DataTypes.TEXT
-  }, {
-    sequelize,
-    modelName: 'blogPost',
-  });
-  return blogPost;
-};
